Revert "Merge branch 'buildaworldnet-shader_pipeline' into shader_pipeline"
This reverts commit 9a70e8a96dc1418c660ce8f0d87cfc8c1652eb9a, reversing
changes made to 124e4ef9c94734283b2eb9ca059246f06a212763.

+++ include/irr/asset/CCPUMesh.h
@@ -52,7 +52,7 @@ class CCPUMesh final : public ICPUMesh
 		//! returns amount of mesh buffers.
 		virtual uint32_t getMeshBufferCount() const override
 		{
-			return static_cast<uint32_t>(MeshBuffers.size());
+			return MeshBuffers.size();
 		}
 
 		//! returns pointer to a mesh buffer
